+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Update dependency io.cucumber:messages up to v26 ((#38)[https://github.com/cucumber/query/pull/38])
+- JavaScript implementation ([#42](https://github.com/cucumber/junit-xml-formatter/pull/42))
+- Update dependency io.cucumber:messages up to v26 ([#38](https://github.com/cucumber/junit-xml-formatter/pull/38))
+
+### Fixed
+- Attempts that have been superseded by retries are now omitted from the report ([#42](https://github.com/cucumber/junit-xml-formatter/pull/42))
 
 ## [0.5.0] - 2024-06-22
 ### Added